+++ CHANGELOG.md
@@ -6,12 +6,13 @@ All notable changes will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased <!-- 1.5.0 -->
+## Unreleased <!-- 2.0.0 - major release, even tho there aren't "breaking" changes (well, adding runtimes that aren't Node to the "F*ckingNODE" project is kinda "breaking") -->
 
 ### Added
 
 - Added an install script for Microsoft Windows.
 - Added the option to flush config files, so the user can save up space.
+- Added partial support for cleanup of both the Bun and Deno JavaScript runtimes.
 
 ### Changed